consumer/requirements.txt
kafka-python-ng==2.2.3
psycopg2-binary==2.9.9
pandas==2.1.2
numpy==1.26.1
xgboost==2.0.3
scikit-learn==1.3.2